requirements.txt
discord.py
psutil
Flask